Time: 1135 ms (5.01%), Space: 191.2 MB (5.10%) - LeetHub

+++ 437-path-sum-iii/437-path-sum-iii.cpp
@@ -18,7 +18,7 @@ public:
         solve(root, t, s,0);
         return ans;
     }
-    void solve(TreeNode* root, int t,unordered_map<long,int> ss,long long cur){
+    void solve(TreeNode* root, int t,unordered_map<long,int> ss,long cur){
         
         if(!root)return;
         // cout<<root->val<<" "<<cur<<endl;
